Arreglado lugar de las funciones
