+++ app/src/main/java/com/artlens/data/facade/Facade.kt
@@ -1,5 +1,6 @@
 package com.artlens.data.facade
 
+import android.content.Context
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import com.artlens.data.models.ArtistResponse
@@ -97,5 +98,9 @@ class Facade(
 
     override suspend fun getArtworkRecommendation(userId: Int): List<ArtworkResponse> {
         return analyticsService.getArtworkRecommendation(userId)
+    }
+
+    override fun downloadFavorites(likedMuseums: List<ArtworkResponse>, context: Context){
+        return artworkService.downloadFavorites(likedMuseums, context)
     }
 }
