+++ app/src/main/java/com/artlens/view/viewmodels/LikesViewModel.kt
@@ -1,11 +1,17 @@
 package com.artlens.view.viewmodels
 
+import android.content.ContentValues
+import android.content.Context
+import android.os.Environment
+import android.provider.MediaStore
+import android.widget.Toast
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.artlens.data.facade.ArtlensFacade
 import com.artlens.data.models.ArtworkResponse
 import com.artlens.utils.UserPreferences
+import java.io.OutputStream
 
 class LikesViewModel(private val facade: ArtlensFacade) : ViewModel() {
     private val _likedMuseums = MutableLiveData<List<ArtworkResponse>>()
@@ -29,6 +35,5 @@ class LikesViewModel(private val facade: ArtlensFacade) : ViewModel() {
             }
         }
     }
-
 
 }
